Cleanup, optimizations & bugfixes : f32 are expensive, some were useless (colors), screen area doesn't fit in a u16, changed it to u32; removed useless conversions in 'random' module, ...
