--- libnw/src/lib.rs
@@ -1,11 +1,0 @@
-#![no_std]
-
-/// math-related utilities
-pub mod libmath {
-    pub use micromath::*;
-}
-
-/// hardware interaction
-pub mod eadk;
-
-pub mod graphics;
