+++ numworks/src/eadk.rs
@@ -61,70 +61,12 @@ pub fn draw_string(
     }
 }
 
-/// Returns a random u32.
-pub fn random() -> u32 {
-    unsafe { eadk_random() }
-}
-
-/// Blocks the thread for a given amount of milliseconds.
-pub fn msleep(ms: u32) {
-    unsafe { eadk_timing_msleep(ms) }
-}
-
-/// Blocks the thread for a given amount of microseconds.
-pub fn usleep(us: u32) {
-    unsafe { eadk_timing_usleep(us) }
-}
-
-/// The number or milliseconds since the start.
-///
-/// The start of what ? I don't know.
-///
-/// Really, nobody knows.
-pub fn millis() -> u64 {
-    unsafe { eadk_timing_millis() }
-}
-
 /// Checks if the usb is plugged.
 ///
 /// The link process will fail if the calculator is not compatible.
 /// You also may need to increase the EADK_API_LEVEL for this to work.
 pub fn usb_is_plugged() -> bool {
     unsafe { eadk_usb_is_plugged() }
-}
-
-/// Checks if the battery is in charge.
-///
-/// The link process will fail if the calculator is not compatible.
-/// You also may need to increase the EADK_API_LEVEL for this to work.
-pub fn battery_is_charging() -> bool {
-    unsafe { eadk_battery_is_charging() }
-}
-
-/// Returns the battery level.
-///
-/// The link process will fail if the calculator is not compatible.
-/// You also may need to increase the EADK_API_LEVEL for this to work.
-pub fn battery_level() -> u8 {
-    unsafe { eadk_battery_level() }
-}
-
-/// Returns the battery voltage.
-///
-/// The link process will fail if the calculator is not compatible.
-/// You also may need to increase the EADK_API_LEVEL for this to work.
-pub fn battery_voltage() -> f32 {
-    unsafe { eadk_battery_voltage() }
-}
-
-/// Sets the screen brightness.
-pub fn backlight_set_brightness(brightness: u8) {
-    unsafe { eadk_backlight_set_brightness(brightness) }
-}
-
-/// Retrieves the screen brightness.
-pub fn backlight_brightness() -> u8 {
-    unsafe { eadk_backlight_brightness() }
 }
 
 /// Retrieves the current state of the keyboard.
@@ -149,11 +91,6 @@ pub fn event_get(timeout: i32) -> EadkKey {
     unsafe { eadk_event_get(&mut timeout as *mut _) }
 }
 
-/// Retrieves the buffer containing the external data needed by this app.
-pub fn external_data() -> &'static [u8] {
-    unsafe { core::slice::from_raw_parts(eadk_external_data, eadk_external_data_size) }
-}
-
 unsafe extern "C" {
     fn eadk_display_push_rect(rect: EadkRect, pixels: *const EadkColor);
     fn eadk_display_push_rect_uniform(rect: EadkRect, color: EadkColor);
@@ -166,20 +103,9 @@ unsafe extern "C" {
         text_color: EadkColor,
         background_color: EadkColor,
     );
-    fn eadk_random() -> u32;
-    fn eadk_timing_msleep(ms: u32);
-    fn eadk_timing_usleep(us: u32);
-    fn eadk_timing_millis() -> u64;
     fn eadk_usb_is_plugged() -> bool;
-    fn eadk_battery_is_charging() -> bool;
-    fn eadk_battery_level() -> u8;
-    fn eadk_battery_voltage() -> f32;
-    fn eadk_backlight_set_brightness(brightness: u8);
-    fn eadk_backlight_brightness() -> u8;
     fn eadk_keyboard_scan() -> u64;
     fn eadk_event_get(timeout: *mut i32) -> EadkKey;
-    static eadk_external_data: *const u8;
-    static eadk_external_data_size: usize;
 }
 
 /// A point on the screen.
